remove deprecated and unused v2 code from network package
