+++ packages/network/src/workers/index.ts
@@ -1,4 +1,3 @@
 export * from "./SyncWorker";
-export * from "./syncUtils";
 export * from "./CacheStore";
 export * from "./constants";
